convert comparitor tests to restAssured, clean up tests, move some methods to util, move classes into respective v1/v2 folders
fixing up the v1 comparator tests, moved shared items into base class, borrow some utility from the other E2E tests

rename tests to E2E'

documentation and cleanup

formatting
